+++ nodalia-cards.manifest.js
@@ -1,23 +1,23 @@
 export default {
-  "pkgVersion": "2.1.1",
-  "contentSha256_12": "c28b078a0d97",
+  "pkgVersion": "2.1.2-alpha.1",
+  "contentSha256_12": "c16875640cc5",
   "file": "nodalia-cards.bundle.js",
   "loaderFile": "nodalia-cards.js",
   "hacsFile": "nodalia-cards.js",
   "compatLoaderFiles": [
-    "nodalia-cards-2.0.5.js",
-    "nodalia-cards-2.1.0.js"
+    "nodalia-cards-2.1.0.js",
+    "nodalia-cards-2.1.1.js"
   ],
-  "splitCoreFile": "nodalia-cards-core-2.1.1.js",
-  "splitCoreSha256_12": "a5716f80a534",
-  "splitSuiteFile": "nodalia-cards-suite-2.1.1.js",
-  "splitSuiteSha256_12": "b5749f77f602",
-  "editorFile": "nodalia-cards-editor-2.1.1.js",
+  "splitCoreFile": "nodalia-cards-core-2.1.2-alpha.1.js",
+  "splitCoreSha256_12": "d6882921f3eb",
+  "splitSuiteFile": "nodalia-cards-suite-2.1.2-alpha.1.js",
+  "splitSuiteSha256_12": "99148ca27fd1",
+  "editorFile": "nodalia-cards-editor-2.1.2-alpha.1.js",
   "editorSha256_12": "7f50f338e82a"
 };
-export const pkgVersion = "2.1.1";
-export const contentSha256_12 = "c28b078a0d97";
+export const pkgVersion = "2.1.2-alpha.1";
+export const contentSha256_12 = "c16875640cc5";
 export const file = "nodalia-cards.bundle.js";
-export const splitCoreFile = "nodalia-cards-core-2.1.1.js";
-export const splitSuiteFile = "nodalia-cards-suite-2.1.1.js";
-export const editorFile = "nodalia-cards-editor-2.1.1.js";
+export const splitCoreFile = "nodalia-cards-core-2.1.2-alpha.1.js";
+export const splitSuiteFile = "nodalia-cards-suite-2.1.2-alpha.1.js";
+export const editorFile = "nodalia-cards-editor-2.1.2-alpha.1.js";